backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
mediapipe
opencv-python
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
requests 